+++ src/common/collections/tenants/checks.js
@@ -1,5 +1,7 @@
 /*
  * pwix:tenants-manager/src/common/collections/tenants/checks.js
+ *
+ * Note: do NOT import Entities nor Records here to prevent any import error (best to leave tenants be fully evaluated before trying to import Entities or Records)
  */
 
 import _ from 'lodash';
@@ -13,9 +15,6 @@ import { ReactiveVar } from 'meteor/reactive-var';
 import { TM } from 'meteor/pwix:typed-message';
 
 import { Tenants } from './index.js';
-
-import { Entities } from '../entities/index.js';
-import { Records } from '../records/index.js';
 
 const logger = Logger.get();
 
@@ -69,26 +68,6 @@ Tenants.checks.contactEmail = async function( value, data, opts ){
             return new TM.TypedMessage({
                 level: TM.MessageLevel.C.ERROR,
                 message: pwixI18n.label( I18N, 'records.check.contact_email_invalid' )
-            });
-        }
-    }
-    return null;
-};
-
-// contact page url
-//  must be valid if set
-Tenants.checks.contactUrl = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.contactUrl()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.contactUrl = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validUrl.isWebUri( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.contact_url_invalid' )
             });
         }
     }
@@ -205,26 +184,6 @@ Tenants.checks.email_row = async function( data, opts ){
     });
 };
 
-// general terms o fuse page url
-//  must be valid if set
-Tenants.checks.gtuUrl = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.gtuUrl()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.gtuUrl = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validUrl.isWebUri( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.gtu_url_invalid' )
-            });
-        }
-    }
-    return null;
-};
-
 // home page url
 //  must be valid if set
 Tenants.checks.homeUrl = async function( value, data, opts ){
@@ -274,39 +233,19 @@ Tenants.checks.label = async function( value, data, opts ){
             });
         };
         const promise =
-            Meteor.isClient ? Meteor.callAsync( 'pwix.TenantsManager.m.Entities.getBy', { label: value }) : Entities.s.getBy({ label: value });
+            Meteor.isClient ? Meteor.callAsync( 'pwix.TenantsManager.m.Entities.getBy', { label: value }) : TenantsManager.Entities.s.getBy({ label: value });
         return promise
             .then(( result ) => {
                 if( result?.length ){
                     return result;
                 } else {
-                    return Meteor.isClient ? Meteor.callAsync( 'pwix.TenantsManager.m.Records.getBy', { label: value }) : Records.s.getBy({ label: value });
+                    return Meteor.isClient ? Meteor.callAsync( 'pwix.TenantsManager.m.Records.getBy', { label: value }) : TenantsManager.Records.s.getBy({ label: value });
                 }
             })
             .then(( result ) => {
                 return fn( result );
             });
     }
-};
-
-// legal terms page url
-//  must be valid if set
-Tenants.checks.legalsUrl = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.legalsUrl()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.legalsUrl = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validUrl.isWebUri( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.legals_url_invalid' )
-            });
-        }
-    }
-    return null;
 };
 
 // logo url
@@ -324,66 +263,6 @@ Tenants.checks.logoUrl = async function( value, data, opts ){
             return new TM.TypedMessage({
                 level: TM.MessageLevel.C.ERROR,
                 message: pwixI18n.label( I18N, 'records.check.logo_url_invalid' )
-            });
-        }
-    }
-    return null;
-};
-
-// personal data management policy page url
-//  must be valid if set
-Tenants.checks.pdmpUrl = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.pdmpUrl()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.pdmpUrl = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validUrl.isWebUri( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.pdmp_url_invalid' )
-            });
-        }
-    }
-    return null;
-};
-
-// support email
-//  must be valid if set
-Tenants.checks.supportEmail = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.supportEmail()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.supportEmail = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validator.validate( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.support_email_invalid' )
-            });
-        }
-    }
-    return null;
-};
-
-// support page url
-//  must be valid if set
-Tenants.checks.supportUrl = async function( value, data, opts ){
-    _assert_data_content( 'Tenants.checks.supportUrl()', data );
-    let item = data.entity.get().DYN.records[data.index].get();
-    if( opts.update !== false ){
-        item.supportUrl = value;
-        data.entity.get().DYN.records[data.index].set( item );
-    }
-    if( value ){
-        if( !validUrl.isWebUri( value )){
-            return new TM.TypedMessage({
-                level: TM.MessageLevel.C.ERROR,
-                message: pwixI18n.label( I18N, 'records.check.support_url_invalid' )
             });
         }
     }